Fixed visual studio compiler warnings
It may take a few more iterations before all warnings are fixed (because I
can't compile on visual studio myself).

Many of these warnings were related to 64->32 bit truncations. Some of those
were bugs that should have been fixed in the earlier 64-bit related patches.
So these warnings are definitely useful.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13095 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/ReverseManager.cc
@@ -299,7 +299,7 @@ void ReverseManager::debugInfo(TclObject& result) const
 	// TODO this is useful during development, but for the end user this
 	// information means nothing. We should remove this later.
 	StringOp::Builder res;
-	unsigned totalSize = 0;
+	size_t totalSize = 0;
 	for (auto& p : history.chunks) {
 		auto& chunk = p.second;
 		res << p.first << ' '
