+++ src/console/OSDText.hh
@@ -36,8 +36,8 @@ private:
 	size_t split(const std::string& line, unsigned maxWidth,
 		FindSplitPointFunc findSplitPoint, CantSplitFunc cantSplit,
 		bool removeTrailingSpaces) const;
-	unsigned splitAtChar(const std::string& line, unsigned maxWidth) const;
-	unsigned splitAtWord(const std::string& line, unsigned maxWidth) const;
+	size_t splitAtChar(const std::string& line, unsigned maxWidth) const;
+	size_t splitAtWord(const std::string& line, unsigned maxWidth) const;
 	std::string getCharWrappedText(const std::string& text, unsigned maxWidth) const;
 	std::string getWordWrappedText(const std::string& text, unsigned maxWidth) const;
