+++ src/sound/SDLSoundDriver.cc
@@ -102,7 +102,7 @@ void SDLSoundDriver::audioCallbackHelper(void* userdata, byte* strm, int len)
 unsigned SDLSoundDriver::getBufferFilled() const
 {
 	int result = writeIdx - readIdx;
-	if (result < 0) result += mixBuffer.size();
+	if (result < 0) result += unsigned(mixBuffer.size());
 	assert((0 <= result) && (unsigned(result) < mixBuffer.size()));
 	return result;
 }
@@ -113,7 +113,7 @@ unsigned SDLSoundDriver::getBufferFree() const
 	// (in both cases readIx would be equal to writeIdx), so instead
 	// we define full as '(writeIdx + 2) == readIdx' (note that index
 	// increases in steps of 2 (stereo)).
-	int result = mixBuffer.size() - 2 - getBufferFilled();
+	int result = unsigned(mixBuffer.size()) - 2 - getBufferFilled();
 	assert((0 <= result) && (unsigned(result) < mixBuffer.size()));
 	return result;
 }
@@ -128,7 +128,7 @@ void SDLSoundDriver::audioCallback(short* stream, unsigned len)
 		memcpy(stream, &mixBuffer[readIdx], num * sizeof(short));
 		readIdx += num;
 	} else {
-		unsigned len1 = mixBuffer.size() - readIdx;
+		unsigned len1 = unsigned(mixBuffer.size()) - readIdx;
 		memcpy(stream, &mixBuffer[readIdx], len1 * sizeof(short));
 		unsigned len2 = num - len1;
 		memcpy(&stream[len1], &mixBuffer[0], len2 * sizeof(short));
@@ -168,7 +168,7 @@ void SDLSoundDriver::uploadBuffer(short* buffer, unsigned len)
 		memcpy(&mixBuffer[writeIdx], buffer, len * sizeof(short));
 		writeIdx += len;
 	} else {
-		unsigned len1 = mixBuffer.size() - writeIdx;
+		unsigned len1 = unsigned(mixBuffer.size()) - writeIdx;
 		memcpy(&mixBuffer[writeIdx], buffer, len1 * sizeof(short));
 		unsigned len2 = len - len1;
 		memcpy(&mixBuffer[0], &buffer[len1], len2 * sizeof(short));
